+++ packages/editor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @i-prikot/editor
 
+## 0.2.2
+
+### Patch Changes
+
+- Ship the editor runtime bundle, stylesheet, and referenced assets in the published package.
+- Updated dependencies
+  - @i-prikot/editor-schema@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes