+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @i-prikot/editor-schema
 
+## 0.2.2
+
+### Patch Changes
+
+- Ship the editor runtime bundle, stylesheet, and referenced assets in the published package.
+
 ## 0.2.1
 
 ### Patch Changes